requirements.txt
ruamel.yaml
pytest
pandas
ebaysdk
slackclient
lxml
luigi
requests
